pyproject.toml
# if poetry update takes forever
# export PYTHON_KEYRING_BACKEND=keyring.backends.fail.Keyring

[tool.poetry]
name = "webtoolkitex"
version = "0.0.11"
description = "Web tools and interfaces for Internet data processing."
authors = ["Iwan Grozny <renegat@renegat0x0.ddns.net>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"
yt-dlp = "^2025.10.22"
beautifulsoup4 = "^4.13.5"
lxml = "^5.3.0"
brutefeedparser="^0.10.5"
pytz = "^2024.2"
psutil="*"
flask="*"
url-cleaner="*"
webtoolkit="^0.0.168"
stealth-requests = "^2.0.4"
curl-cffi = "^0.14.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
